requirements.txt
Flask==2.3.3
Flask-JWT-Simple==0.0.3
psycopg2-binary==2.9.9
